lattice_signer_service/requirements.txt
Flask==2.0.3
pycryptodome==3.10.1
requests==2.25.1
